+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _Unreleased_
 * Compass was hidden when location tracking was enabled.
 * QR code scanner removes tracking parameters.
 * Added a button to edit an address in the editor pane.
+* Disabled editing addresses on buildings in Italy and France.
 
 ## 7.0